Fix implementation errors

+++ app/Console/Commands/GenerateThumbnails.php
@@ -93,7 +93,7 @@ class GenerateThumbnails extends Command
         $path = escapeshellarg($video->file_path);
         $duration = shell_exec("ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 $path 2>/dev/null");
         if ($duration) {
-            $duration = trim($duration);
+            $duration = (float)trim($duration);
         }
 
         // Seek to 30% of video duration, or 10 seconds if duration is unknown.
